rm renderingdirmanager
