+++ apps/blender/task/blenderrendertask.py
@@ -14,7 +14,7 @@ from apps.blender.blenderenvironment import BlenderEnvironment
 from apps.blender.resources.scenefileeditor import regenerate_blender_crop_file
 from apps.rendering.resources.renderingtaskcollector import RenderingTaskCollector, exr_to_pil
 from apps.rendering.task.framerenderingtask import FrameRenderingTask, FrameRenderingTaskBuilder
-from apps.rendering.task.renderingdirmanager import get_test_task_path, find_task_script
+from golem.resource.dirmanager import get_test_task_path, find_task_script
 from apps.rendering.task.renderingtask import AcceptClientVerdict
 from apps.rendering.task.renderingtaskstate import RendererDefaults, RendererInfo
