+++ golem/task/localcomputer.py
@@ -4,10 +4,9 @@ import shutil
 from threading import Lock
 
 from golem.docker.task_thread import DockerTaskThread
-from golem.task.taskbase import Task, resource_types
+from golem.resource.dirmanager import get_test_task_path, get_test_task_tmp_path
 from golem.resource.resource import TaskResourceHeader, decompress_dir
-
-from apps.rendering.task.renderingdirmanager import get_test_task_path, get_test_task_tmp_path
+from golem.task.taskbase import Task, resource_types
 
 logger = logging.getLogger("golem.task")
