+++ tests/golem/docker/test_luxrender_job.py
@@ -2,8 +2,8 @@ import os
 from os import path
 import shutil
 
-from apps.rendering.task.renderingdirmanager import find_task_script
 from golem.core.common import get_golem_path
+from golem.resource.dirmanager import find_task_script
 
 from test_docker_job import TestDockerJob
