+++ tests/golem/resource/test_dirmanager.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-from golem.resource.dirmanager import DirManager
+from golem.resource.dirmanager import DirManager, find_task_script
 from golem.tools.testdirfixture import TestDirFixture
 
 
@@ -174,3 +174,15 @@ class TestDirManager(TestDirFixture):
         self.assertFalse(os.path.isfile(file2))
         self.assertFalse(os.path.isfile(file3))
         self.assertFalse(os.path.isdir(dir1))
+
+
+class TestFindTaskScript(TestDirFixture):
+    def test_find_task_script(self):
+        script_path = os.path.join(self.path, "resources", "scripts")
+        os.makedirs(script_path)
+        script = os.path.join(script_path, "bla")
+        open(script, "w").close()
+        task_file = os.path.join(self.path, "task", "testtask.py")
+        path = find_task_script(self.path, "bla")
+        self.assertTrue(os.path.isdir(os.path.dirname(path)))
+        self.assertEqual(os.path.basename(path), "bla")
